Persist interrupted steps as intermediate runs
This is similar to how we previously changed it to also persist the final run. But we need to tweak the logic for showing these intermediate runs. It's quite tricky to balance this so we 1) don't miss any runs or user responses 2) don't end up with double runs by having both intermediate and non-intermediate runs 3) can still merge the runs that belong together when we filter out the non-intermediate runs.