Allow to reset optional default model parameters
Without this change, it would be impossible to ever set the default seed back to undefined once you have set it to something else. The partial type used here to cover both default model and default parameters is either elegant or a bit of a hack depending on your perspective, but in practice we either set the model or the rest of the parameters, so it should be safe to take this shortcut.