Make pre-merge migrations safe to run after push
Once we run the post-merge migration, the old unencrypted apiKey attribute will be cleared. If we then run the pre-migration again (which should only happen in dev but could easily happen there if we add another migration before merging to prod), we should retain the encrypted keys and not reset them or we will have lost our API keys which is scary.