Add dropdown to select dynamic interrupt mode
We will only show this when there are non-empty values for dynamic input variables. In that case you will be able to use them as either suggestions for manual responses (currently the only option), fixed mocked responses (previously the only option when we couldn't test interrupts in the UI), or prompts for automatic responses (the new option we want to prototype).