Add max responses in test config
If we're going to generate responses automatically we need to be able to set a limit so the conversation doesn't go on forever (we also need this in the case of fixed mocked responses which we are reintroducing). Having this extra parameter probably means we can get by with a simple boolean for autoRespond though (max responses zero means manual mode regardless of whether autoRespond is true/false/undefined).