Send dynamic inputs separately to run version
The problem with simply including the dynamic inputs is that the code interrupts might never trigger because the value is already present in the inputs. We can split them up relatively easy in the RunButtons callback logic, but note that we don't include the dynamic inputs when running a manual continuation in RunTimeline (which is probably a limitation we can live with for now, i.e. once you reach a manual interrupt, you need to run the consecutive interrupts manually as well).