Keep partial run offset separately from index
We need the offset to ensure each partial run gets a unique ID in the case of auto responses (or soon: loops). These temporary IDs are also used to sort the partial runs. But we want to keep the index of the partial run reflect the node in the chain, otherwise things like highlighting the active node won't work anymore.