Bump timestamp when singing up user again
If a user tries to sign up again, we will bump the createdAt timestamp (but not when we give a user access that previously signed up). The potential downside is that we would no longer be able to tell who signed up earliest, i.e. signing up again would move you to the "end" of the waitlist, but we can worry about that (and change it again) if we ever end up in a situation where that matters (right now it matters more to be able to tell recent activity).