Update node editor state on chain editor changes
Currently, if you click a node to open its editor, you can still make changes to the entire chain in the left view as well, until you actually start making changes to the node (e.g. edit the code of the code block). These chain changes occur when adding or removing an item from the chain (which we already considered before), but also when interacting with the output mapper or the include context checkbox. In all these cases, we should actually update the start state for the node editor, and it is also safe to do that, because at this point no changes have been made to the node yet (otherwise the chain editor would be disabled). Sample edge case that this fixes: you click on a code block, change something in an output mapper, and hit Save and close (either with or without making changes to the code). The chain changes should not be overwritten by any node changes (or no node changes).