Persist project on redirecting to install GitHub app
This is definitely a bit hacky (i.e. storing the project ID in the environment of a github provider entity with user scope which won't be otherwise used). But we will need to know what project to store the actual github provider for once the user has installed the app. Having this endpoint does mean we don't need to expose the app link client side (although it isn't really sensitive).