Implement GitHub callback
So here we get a user authorisation code and installation ID. We need to verify that the installation ID is verified to the user (otherwise it could be spoofed and we might end up reading from or writing to the wrong repository). This requires first getting an access token. But once we have that, we can also fetch the repositories for that installation and we store them in the environment for the provider (while the installation ID is stored in the api key).