Implement GitHub project settings flow
This was harder than expected. Still feels a bit hacky (e.g. the way we propagate the installation ID from user scope to project scope on the backend and the way we try to parse the repositories from the available provider on the client) but the alternative would require adding dedicated endpoints.