Add cleanup datastore
The idea is that we delete the parent keys immediately, but then also store them so we can delete associated entities asynchronously through a cron job (e.g. in case there are many logs for an endpoint, we don't need to block while deleting all of them but can safely do it later).