Clean up associated entities for workspace
This one is a bit special because we do want to delete the project entities belonging to the workspace straight away (and schedule their associated entities for cleanup). Should be safe to assume that there won't be workspaces with more than 1000 projects any time soon (and even if there are, they will still get cleaned up with the next cron job).