Ignore orphaned log entries
Due to the cleanup cron job, it's possible that we still load log entries for endpoints that have already been deleted (until the job next runs again). Hence we should filter these out on the client side (in theory we could filter them out server side as well but it's a bit tricky because we might fetch them separately from the endpoints and doesn't feel worth adding too much additional logic for this).