Avoid capturing original input values
Not sure why this didn't repro with the old editor but with the new editor it was easy to get into a situation where it keeps calling the API when switching between values that have already been synced. We should make sure we use the current value of both original and modified input values here.