Avoid capturing old config in prompt update
The reason we need to be more careful about this than with the previous editor, is that we don't update the editor when the setValue callback changes but rather configure it upfront (there may be an alternative way of doing this but it doesn't feel wrong to be more explicit in the update functions like this). The other tricky thing is that we need to schedule the setModifiedVersion call to avoid updating a parent component while updating the child component.