Update active prompt version when node changes
This fixes an ugly bug that we've probably had for a long time, where, if you click on one prompt in the chain and then another, it marks the editor as dirty even if you didn't change anything. It became more obvious now that the new editor sends a dummy change event on initialisation.