Refresh active prompt node on save chain
When you save a prompt node in a chain, the version of that prompt in use by that chain will often change, so we should refresh it in the prompt cache after saving the chain. This is tricky because it also requires the latest state of the project (to resolve the references) rather than capturing the previous state. Easiest way to do this is to refresh the project and then clear the prompt from the cache which will trigger a reload with the new project.