Don't refresh prompt twice on save in chain
We need to split up this saveAndRefreshPrompt. When you select a different version, we should trigger a prompt save in case you made any changes (so a draft version might be created or updated) but we don't need to save the chain yet. If you save and close a prompt node, we should save the prompt and use the resulting version to update the chain, but in this case we don't need to refresh the prompt as this will be taken care of by the chain view (which will first refresh the chain and the project because the referenced items may change).