Don't select version on inserting new prompt
When we insert a new prompt, we should not select its initial version yet, because we don't select a version when you add an existing prompt either (unless it would already have been loaded in the prompt cache). For now we can assume that the user will click Save and Close after inserting a prompt (and either selecting a version or modifying the selected version or just confirm to use the selected version). We could improve this e.g. by removing the inserted node on hitting Cancel but we won't bother with that edge case now. What this is fixes is that you don't get an extra unused empty version when you insert a new prompt, edit it and then hit Save and close.