Consider relevant part of empty row in test data
We need to check if the row is empty when deciding to add a new row or delete one, but in the toggle logic, we should only consider the part of the row with relevant variables (this was broken before the changes to the test data today e.g. you unselect all did not work with sparse test data).