Add edge case in global popup location
Normally, when we fully specify the location, we treat it as absolute and don't touch it. But in this case, we want to specify left and right but still centre vertically. So we add a case where if top and bottom are zero we treat them as such if the content fills the entire height but otherwise centre vertically. Feels a bit hacky though (more heavy alternative would be to add more explicit options like centerVertically to the location parameter or a callback for clients to specify the location based on parent and child).