Fix padding when scrolling expanded test data
Removing the top padding from the button and adding the bottom padding in the expanded child class instead doesn't work well when we need to scroll to reveal all content in the expanded test data pane. So instead we move it a bit closer when collapsed by tweaking the margin/padding.