Re-enable edge detection for hovering sidebar
We're making the edge very narrow though to avoid conflicting with e.g. buttons to expand test data or prompt sections. Having a narrow edge is probably fine especially when your window is to the edge of your screen in which case you can just move your mouse as far left as possible to immediately trigger the menu.