Ensure document and editor selection are same
There was an issue where, e.g. if you selected a test data value while you also had a prompt input on screen, it would extract the editor selection in the prompt input and combine it with the document selection in the test data contentEditable to show a Create or Remove Input popup in the test data editor.