Move expand icon in last column a bit to the left
So it doesn't conflict with a potential scroll bar. This is redundant when there is no scrollbar not too bad either as it horizontally aligns with the same icon to expand the entire test data pane (which we should probably add in the chain UI as well at some point).