Expose dummy version when no active draft version
If you start making changes without there being a persisted draft version, we want to expose a dummy version (think of it as a ghost draft version). But we need to take care elsewhere because this version doesn't exist on the backend yet, so e.g. you can't comment on it yet. Also the selection behaviour in the timeline needs some special casing (depending on context, the "active version" is either the dummy version or its persisted source version, e.g. when saving we need both).