Reset current version when draft version detected
This is needed because the active version won't change when you navigate from the dummy version to its source version. In that case, the dummy version will be persisted, but once that has happened, we need to reset the current version. It all gets a bit tricky so I'm still not sure this dummy version is a good idea to begin with.