Validate prompt config on load from user preset
Otherwise you might end up with e.g. a new prompt that has a PaLM model but also has an OpenAI seed parameter set (as would be revealed if you switched the model to GPT after creating the new prompt). We should apply the same logic we apply when changing the config in the UI.