Allow to abort stream reader when running version
By the time we have a partial run to show the cancel button on, the fetch has already returned a stream reader which we are consuming, so aborting won't actually abort the fetch but rather the stream reading. Obviously this does not cancel the server side processing yet (we'll need to detect that and process it separately).