Abort server side processing on request closed
When we cancel a request to run a version on the client, we also want to stop processing on the server. Apparently the way to do this is to detect that the request has been closed with a listener and then we can pass the abort controller around. Ideally we also want to cancel the LLM request to avoid paying for unneeded tokens or hitting rate limits, but only OpenAI and Anthropic appear to support this properly for now.