Also save input values on single empty input
At first I thought this was by design, but it doesn't really make sense to save when there are two empty row values but not when there is a single empty row value. The logic that prevents us from creating a new column when editing a prompt variable lives elsewhere (it can't rely on this or it wouldn't work if a variable had two empty row values).