Pass in isFirst/isLast rather than leftBorder
Better to have the header decide its own markup based on semantics, and this will also allow to render popup button a bit more to the left for the last header item. Also fixes an issue in the test data value popup where leftBorder wasn't passed in even though there was a gutter column (now the default is "not first" rather than "not leftBorder" i.e. first).