Fix inconsistent semantics for empty table state
It's actually better to make this consistent. You could argue that the downside is that now, even when you have just started from scratch, you can already reset or save the data (because it is no longer considered empty) and that doesn't really make sense, but actually it does make sense because otherwise you won't be able to go back easily to the empty state where you import CSV. There is in fact a difference between empty state and new table, so we should not treat the latter as empty.