Implement cancellation through api endpoint
It looks like being able to detect when requests close is going to be infeasible on GCP because the connections are kept alive. So instead we will send a cancel request from the client. The problem with is that we somehow need to map the cancel request to the original request to be cancelled within a stateless HTTP server. The session_affinity option should help with this but it feels a bit hacky (and we will only know if it works at all after pushing to dev).