Expose project owners before non-owners
We already expose project members before workspace members, so we might as well expose the project owners before that. This ensures that the first project member you see listed is always a project owner (e.g. in case you need to reach out for permissions), without having to add much weight to it in the UI.