Fix some styling issues with single editor cell
We can actually simplify some of the styling conditions because we are splitting up active and inactive cells. Note that we need to use onMouseDown rather than onClick on the inactive cell because the onBlur on the previously active cell may cause the row height to change (line-clamp) so the onClick on say the next row might no longer be on the next row by the time it completes.