Simplify deselect cell logic
We no longer need to worry about onBlur being called after onFocus (see 199c16e), because the editor component will be unloaded if another cell is activated. The only case where onBlur is called is if we in fact lose focus for the entire table, which is exactly when we want to deactivate the previously active cell.