Move cursor to mouse location on editor load
Because the editor won't be loaded yet when we click on a table cell, it also won't be able to set its cursor to where the user clicked within the cell unless we pass in that location. This isn't perfect because the line clamping might make the rows jump, but it's better than nothing.