Persist changes on activating cell
The onBlur event is no longer called when switching cells because the entire editor component for the previously active cell gets unloaded. So we need to check for changes not only when deactivating a cell (which only gets triggered when the entire table loses focus) but also when activating another cell.