Shorten final connector in single branch chain
When there are multiple branches, there is a horizontal line at the very bottom linking them back to the Output node (but skipping the looping branches). In that case, we need a little extra downstroke before the horizontal line to avoid the horizontal line touching the final add button in the column. But if there is only a single branch (or additional branches are all looping), we don't want this.