Make ActiveItemCache generic
This forces us to be a bit more specific between ActiveItemCache<ActivePrompt | ActiveChain>, ActiveItemCache<ActivePrompt> and ChainItemCache, but that's actually a good thing, i.e. it gets rid of most "as ActivePrompt" casts and instead you get a cast like the one on line 84 in useDiffContent.tsx which makes the boundary more clear.