Update gcloud instructions in readme
If you are using multiple accounts with Google Cloud, it is not enough to run "gcloud auth login" and "gcloud init" to switch between accounts and projects, but you need to run this instead to update the application default credentials that will be used e.g. when calling the remote datastore.