Don't allow overlapping comments
Responses are currently split up in spans so we can render the comments differently. Previously, we would allow starting a new comment across two spans (e.g. starting selection within an existing comment and ending in the text beyond the comment). When the selection started or ended in a previous comment, we would end up mapping the quoted range to the previous comment, but you could still create a new comment that both starts before and ends after a previous comment. We can probably simply avoid these edge cases instead (and probably should before we start messing with markdown formatting).