Create inline comment spans after load
This is quite tricky but important if we want to be able to render markdown with inline comments. Instead of generating the spans during rendering, we will just render the whole output, and then surround the comment ranges with spans. The reason it's tricky is because there is no simple mapping between the raw start index of the comment in the output string and the start offset of its range in the rendered node (because you are dealing with potentially a complex tree of nodes). The approach taken should be relatively robust in the sense that the startIndex is still the raw index and independent of the actual rendering (i.e. it should not change when we switch to markdown rendering). But it is a bit expensive because we need to process the whole node tree to map the startIndex to the actual node and range indices.