Don't render markdown with legacy inline comments
This is hacky. The alternative would be to consider the number of whitespace characters that are being trimmed by react-markdown by dynamically comparing the raw output with the textContent in the nodes while processing the tree This doesn't feel worth the trouble given there are only a handful of inline comments in prod, but we should probably at least make sure those comments still show up at the right place, which we can do by simply not rendering markdown in that case.