List workspace owners before other members
This is similar to what we do for projects (getProjectAndWorkspaceUsers), i.e. we don't expose all member roles to non-owners, but we do list the owners first so anyone can identify the first user listed as being an owner (e.g. in case you need to reach out about permissions).