Change permissions for leaving/deleting workspace
Previously, there was no way to change ownership of a workspace. So the logic was that you could only delete a workspace if there were no other members in it and otherwise you could leave it (so even though the owner would never change, they could leave the project and we would allow whoever remained last to delete the workspace without checking for ownership at that point). But now that we will expose workspace roles and allow to change them, it makes more sense to change this logic so now all owners can delete the workspace and you can leave the project unless you are the only owner (in which case you should promote some other member to owner first).