Add option to make modal dialog non-dismissable
We already have a cancellable option, but if that is false, it just means don't show the cancel button, and you can still dismiss the dialog by clicking outside of it or by pressing Escape. So we need to add this option to make it proper modal (which is rarely needed).