Get rid of "free" model promotion
We no longer have Google Cloud credits so we should no longer promote this so heavily. Whoever ends up running the instance of PlayFetch on their Google Cloud infra will end up paying for the use of these models (with the current implementation), so tweaked the language a bit to reflect this.